+++ src/extractors/authenticated_user.rs
@@ -1,0 +1,41 @@
+use axum::{extract::FromRequestParts, http::request::Parts};
+use serde::{Deserialize, Serialize};
+use tower_sessions::Session;
+
+use crate::{error::AppError, handlers::auth::ZauthUser};
+
+#[derive(Debug, Serialize, Deserialize, PartialEq)]
+pub struct AuthenticatedUser {
+    pub id: u32,
+    pub username: String,
+    pub admin: bool,
+}
+
+impl<S> FromRequestParts<S> for AuthenticatedUser
+where
+    S: Send + Sync,
+{
+    type Rejection = AppError;
+
+    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
+        let session = Session::from_request_parts(parts, _state)
+            .await
+            .map_err(|(_, msg)| AppError::Internal(msg.into()))?;
+
+        let user: Option<AuthenticatedUser> =
+            session.get("user").await.map_err(AppError::Session)?;
+        user.ok_or(AppError::NotLoggedIn)
+    }
+}
+
+impl From<ZauthUser> for AuthenticatedUser {
+    fn from(user: ZauthUser) -> Self {
+        let admin =
+            user.roles.contains(&"bestuur".into()) || user.roles.contains(&"zpi_admin".into());
+        Self {
+            id: user.id,
+            username: user.username,
+            admin,
+        }
+    }
+}
