+++ src/extractors/mod.rs
@@ -1,0 +1,7 @@
+pub mod admin;
+pub mod authenticated_user;
+pub mod config;
+pub mod database;
+
+pub use admin::Admin;
+pub use authenticated_user::AuthenticatedUser;
